Add a comprehensive compilation guide for the NewCity project and refactor various source files for improved type safety and clarity.
- Introduced `COMPILATION_GUIDE.md` with detailed instructions for compiling on Linux, including prerequisites, build options, and development workflow.
- Refactored multiple source files to cast enumerations to integers for better type safety and consistency.
- Updated variable names for clarity and improved readability in several files, including `activity.cpp`, `business.cpp`, and `vehicle.cpp`.
- Enhanced logging and error handling in various components to ensure better debugging and maintenance.
- Upgraded STB's library from 1.16 to 1.22

+++ src/activity.cpp
@@ -146,7 +146,7 @@ float scoreInterview(item ndx) {
   score += c(CInterviewLightFactor)*getLightLevel();
   float timeSinceWork = getCurrentDateTime() - f->lastWorkTime;
   score += timeSinceWork * c(CInterviewTimeSince);
-  score += getEducationForPerson(ndx) * c(CInterviewEducation);
+  score += (int)getEducationForPerson(ndx) * c(CInterviewEducation);
   if (person->activity == InterviewActivity) {
     score += activityInertia(person, CInterviewInertiaA,
         CInterviewInertiaB, CInterviewInertiaC, energy);
